# api/public.openapi.yaml
openapi: "3.1.0"
info:
  title: Auth Service — Public API
  version: 0.1.0
  description: >
    Authentication and authorization endpoints for the QuantFlow Studio ecosystem.
    Serves Users (humans) and Systems (services + AI agents).


    OIDC provider config: the OIDC authorization-code flow (`/oauth/authorize`,
    `/oauth/token`, `/userinfo`, `/.well-known/openid-configuration`) requires
    `OIDC_LOGIN_UI_URL` to be set — `/oauth/authorize` returns an internal
    error otherwise, since it has nowhere to redirect the end user to
    authenticate. `OIDC_CONSENT_UI_URL` is optional and defaults to
    `OIDC_LOGIN_UI_URL` when unset. `OIDC_ISSUER_URL` sets both the `issuer`
    field of the discovery document and the `iss` claim on issued access and
    ID tokens (GH-468) — prior to that fix, `iss` was hardcoded and
    `OIDC_ISSUER_URL` had no effect on issued tokens; deployments that already
    set a non-default `OIDC_ISSUER_URL` will see `iss` change to match it.
  contact:
    name: QuantFlow Studio
    url: https://quantflow.studio
  license:
    name: Proprietary
servers:
  - url: https://auth.quantflow.studio
    description: Production
  - url: http://localhost:4000
    description: Local development

paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      tags: [Health]
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /auth/register:
    post:
      operationId: register
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Email already registered

  /auth/login:
    post:
      operationId: login
      summary: Authenticate with email and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/token:
    post:
      operationId: token
      summary: Exchange refresh token or client credentials for a new access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: Token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/revoke:
    post:
      operationId: revokeToken
      summary: Revoke a refresh token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeRequest"
      responses:
        "204":
          description: Token revoked
        "400":
          $ref: "#/components/responses/ValidationError"

  /auth/password/reset:
    post:
      operationId: resetPassword
      summary: Request a password reset
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "202":
          description: Reset email sent (if account exists)

  /auth/password/reset/confirm:
    post:
      operationId: confirmPasswordReset
      summary: Confirm password reset with token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmRequest"
      responses:
        "204":
          description: Password updated
        "400":
          $ref: "#/components/responses/ValidationError"

  /auth/me:
    get:
      operationId: getMe
      summary: Get current user info
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      operationId: updateProfile
      summary: Update current user's display name
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/me/password:
    put:
      operationId: changePassword
      summary: Change current user password
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordChangeRequest"
      responses:
        "204":
          description: Password changed
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/me/email:
    post:
      operationId: requestEmailChange
      summary: Request a change of the current user's account email
      tags: [Auth]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailChangeRequest"
      responses:
        "202":
          description: Email change pending confirmation from the new address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailChangeResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  /auth/email-change/confirm:
    post:
      operationId: confirmEmailChange
      summary: Confirm an account email change from the new address
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailChangeConfirmRequest"
      responses:
        "200":
          description: Email changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailChangeStatusResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"

  /auth/email-change/revert:
    post:
      operationId: revertEmailChange
      summary: Revert an account email change from the original address
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailChangeRevertRequest"
      responses:
        "200":
          description: Email change reverted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailChangeStatusResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
        "410":
          $ref: "#/components/responses/Gone"

  /auth/logout:
    post:
      operationId: logout
      summary: Logout current session
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Logged out
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/logout/all:
    post:
      operationId: logoutAll
      summary: Logout all sessions
      tags: [Auth]
      security:
        - bearerAuth: []
      responses:
        "204":
          description: All sessions terminated
        "401":
          $ref: "#/components/responses/Unauthorized"

  /.well-known/jwks.json:
    get:
      operationId: getJWKS
      summary: JSON Web Key Set
      tags: [Discovery]
      responses:
        "200":
          description: Public keys for token verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JWKSResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Request understood but not permitted (e.g. wrong password)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with existing state (e.g. email already in use)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Gone:
      description: The referenced token has expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Feature not currently configured
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
      required: [status]

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 15
        name:
          type: string
          minLength: 1
          maxLength: 255
      required: [email, password, name]

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]

    TokenRequest:
      type: object
      properties:
        grant_type:
          type: string
          enum: [refresh_token, client_credentials]
        refresh_token:
          type: string
        client_id:
          type: string
        client_secret:
          type: string
      required: [grant_type]

    RevokeRequest:
      type: object
      properties:
        token:
          type: string
      required: [token]

    PasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]

    PasswordResetConfirmRequest:
      type: object
      properties:
        token:
          type: string
        new_password:
          type: string
          minLength: 15
      required: [token, new_password]

    PasswordChangeRequest:
      type: object
      properties:
        old_password:
          type: string
        new_password:
          type: string
          minLength: 15
      required: [old_password, new_password]

    ProfileUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
      required: [name]

    EmailChangeRequest:
      type: object
      properties:
        new_email:
          type: string
          format: email
        password:
          type: string
      required: [new_email, password]

    EmailChangeResult:
      type: object
      properties:
        status:
          type: string
          example: pending
        pending_email:
          type: string
          format: email
      required: [status, pending_email]

    EmailChangeConfirmRequest:
      type: object
      properties:
        token:
          type: string
      required: [token]

    EmailChangeRevertRequest:
      type: object
      properties:
        token:
          type: string
      required: [token]

    EmailChangeStatusResponse:
      type: object
      properties:
        status:
          type: string
          example: changed
      required: [status]

    UserInfo:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
      required: [id, email, name]

    AuthResult:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
      required: [access_token, token_type, expires_in]

    JWKSResponse:
      type: object
      properties:
        keys:
          type: array
          items:
            type: object

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
      required: [error, code]